[IE][VPU]: Moves CompilationConfig out of GT header
Keeping config in a separate header simplifies migration
to new interface.
