+++ inference-engine/src/vpu/graph_transformer/include/vpu/graph_transformer.hpp
@@ -21,107 +21,16 @@
 #include <vpu/utils/perf_report.hpp>
 #include <vpu/utils/logger.hpp>
 #include <vpu/utils/optional.hpp>
+#include <vpu/configuration/configuration.hpp>
 
 namespace vpu {
 
 namespace ie = InferenceEngine;
 
-//
-// CompilationConfig
-//
-
 VPU_DECLARE_ENUM(Platform,
     MYRIAD_2 = 2450,
     MYRIAD_X = 2480,
 )
-
-struct CompilationConfig final {
-    //
-    // Compilation options
-    //
-
-    int numSHAVEs = -1;
-    int numCMXSlices = -1;
-    int numExecutors = -1;
-    int tilingCMXLimitKB = -1;
-
-    bool hwOptimization = true;
-    bool hwExtraSplit = false;
-
-    std::string irWithVpuScalesDir;
-
-    std::string customLayers;
-
-    bool detectBatch = true;
-
-    Optional<bool> copyOptimization;
-    Optional<bool> injectSwOps;
-    Optional<bool> packDataInCmx;
-    bool mergeHwPoolToConv = true;
-    bool hwDilation = false;
-    bool forceDeprecatedCnnConversion = false;
-    bool enableEarlyEltwiseReLUFusion = true;
-
-    std::map<std::string, std::vector<int>> ioStrides;
-
-    //
-    // Debug options
-    //
-
-    ie::details::caseless_set<std::string> hwWhiteList;
-    ie::details::caseless_set<std::string> hwBlackList;
-
-    bool hwDisabled(const std::string& layerName) const {
-        if (!hwWhiteList.empty()) {
-            return hwWhiteList.count(layerName) == 0;
-        }
-
-        if (!hwBlackList.empty()) {
-            return hwBlackList.count(layerName) != 0;
-        }
-
-        return false;
-    }
-
-    ie::details::caseless_set<std::string> noneLayers;
-
-    bool skipAllLayers() const {
-        if (noneLayers.size() == 1) {
-            const auto& val = *noneLayers.begin();
-            return val == "*";
-        }
-        return false;
-    }
-
-    bool skipLayerType(const std::string& layerType) const {
-        return noneLayers.count(layerType) != 0;
-    }
-    bool ignoreUnknownLayers = false;
-
-    std::string dumpInternalGraphFileName;
-    std::string dumpInternalGraphDirectory;
-    bool dumpAllPasses;
-
-    bool disableReorder = false;  // TODO: rename to enableReorder and switch logic.
-    bool disableConvertStages = false;
-    bool enablePermuteMerging = true;
-    bool enableReplWithSCRelu = false;
-    bool enableReplaceWithReduceMean = true;
-    bool enableTensorIteratorUnrolling = false;
-    bool forcePureTensorIterator = false;
-    bool enableMemoryTypesAnnotation = false;
-    bool enableWeightsAnalysis = true;
-    bool checkPreprocessingInsideModel = true;
-    bool enableCustomReshapeParam = false;
-
-    //
-    // Deprecated options
-    //
-
-    float inputScale = 1.0f;
-    float inputBias = 0.0f;
-};
-
 
 //
 // DataInfo
